+++ internal/entity/weather.go
@@ -6,9 +6,9 @@ import (
 )
 
 type Weather struct {
-	TempC float64 `json:"temp_C"`
-	TempF float64 `json:"temp_F"`
-	TempK float64 `json:"temp_K"`
+	TempC float32 `json:"temp_C"`
+	TempF float32 `json:"temp_F"`
+	TempK float32 `json:"temp_K"`
 }
 
 func (weather *Weather) CalculateFarenheit() {
